Contrain maximal overlap of read with indel.
This is more fair that using the centerpoint of the indel when investigating
overlapping variants, because the mapper is biased towards alignments with few
edits (i.e. those supporting the reference allele). If a large part of the read
overlaps with the indel, it can lead to a bad mapping quality, or even not mapping
the read at all.