Do not normalize indel likelihoods for ref and alt. Only consider a small window around the indel.
The small window avoids biases due to close variants that would occur because we omit the normalization of ref and alt likelihoods.
We have to omit the normalization because it causes false positives if e.g. a deletion is overlapped by an insertion:
suddenly, a read supporting the insertion looks like a read supporting the deletion because it maps slightly better to the deletion
than to the reference (but bad to both). However, normalization would amplify this difference such that prob_alt becomes very high.
Without normalization, such a read does contribute neither to ref nor to alt, which is what we want.