tolerate reads with SNV position deleted or skipped
`None` is now a valid return value for prob_snv(), as the SNV position
of interest can be deleted (Cigar op 'D') or skipped (Cigar op 'N') in a
read spanning that SNV position, which thus doesn't provide any evidence
and shouldn't be added to the observations in extract_observations(). To
avoid complicating read_observation(), SNV read extraction is taken out
of that function. Tests for previous prob_read_snv in model/sample.rs
are now moved to model/evidence/reads.rs to test prob_snv().