move from rustc-serialize dep to csv 1.0.0-beta.5
This mirrors an update in rust-bio 0.15 compared to rust-bio 0.14.
While this branch compiles properly, 4 tests fail with a
FloatParsingError on field 8 of the first record in the observation csv
files example[2347].obs.txt in the new csv function `deserialize()`. Any
ideas what might have changed in comparison to the `decode()` function
in csv 0.15.0?