Actually remove alleles and calls from the bcf when controlling FDR.
This is the only way to filter here, since there are no allele specific filters in the BCF standard, hence just marking alleles as filtered out does not work. Since filtering is super fast, it should be no issue to create multiple for multiple desired FDRs. The previous approach, which just generated a threshold table for manual filtering is dangerous, since the user would need to ensure that also vartypes are filtered in exactly the same way. Now, we have a nice all-in-one for filtering.