Improved MAPQ handling, that should reduce a bunch of artifacts.
a) For paired end evidence, take the worst MAPQ instead of multiplying them. Rationale: they are not indepently chosen by the mapper. Taking the weaker is the conservative choice.
b) For SNVs and MNVs take the average MAPQ over the entire pileup. Rationale: MAPQ is driven by a) locus ambiguity and b) noise generated by sequencing errors and variants at homologous loci. We want to get rid of b), since it can introduce artifacts if it happens to artificially enlarge a MAPQ for FP-variant supporting reads. Hence, we reduce the MAPQ to the average, and conservatively keep MAPQs lower than the average.