feat: overhauled somatic prior (#209)
The effective mutation rate now defines the expected fraction of sites with somatic VAF > 0.0. The a priori expectation of the height of the VAF is now uniformly distributed. This improves better resembles intuition when setting the prior assumptions. Further, it avoids cases where higher somatic VAFs becomes so unlikely that suddenly the model assumed germline mutations even against the evidence in e.g. a normal sample.