feat: add MATEID based recognition of connected breakends (#313)
Before, only EVENT tags where supported. Now, also MATEID without an
EVENT tag is enough to construct a pair of connected breakends.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation at
https://github.com/varlociraptor/varlociraptor.github.io is updated in a
separate PR to reflect the changes or this is not necessary (e.g. if the
change does neither modify the calling grammar nor the behavior or
functionalities of Varlociraptor).