fix: match EF-tag (#337)
### Description

When checking for the presence of the `EF`-tag, it is never found
because rust-htslib interprets it as an unsigned integer.
For the tag being recognized I have added the U8 to U32 enums. 
As I am not aware if rust-htslib may interpret the tag as a signed
integer I left the corresponding enums as they where.
### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation at
https://github.com/varlociraptor/varlociraptor.github.io is updated in a
separate PR to reflect the changes or this is not necessary (e.g. if the
change does neither modify the calling grammar nor the behavior or
functionalities of Varlociraptor).