feat: by default, use prior to guard absence or presence of variant but weight all decisions about other events just by the observations (#349)
This results in more intuitive classifications, and also avoids pitfalls
generated by very low mutation rates.

An example for such a pitfall is the case where the somatic mutation
rate is set much lower in the normal sample than in a tumor sample. In
such cases, it could have happened that Varlociraptor puts aside the
evidence for the variant being somatic in the normal sample just because
it is cheaper to consider it a tumor sample only variant. While this is
statistically correct given the prior beliefs in that case, it makes the
model very dependent on those priors being chosen correctly, which is in
practice very difficult.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation at
https://github.com/varlociraptor/varlociraptor.github.io is updated in a
separate PR to reflect the changes or this is not necessary (e.g. if the
change does neither modify the calling grammar nor the behavior or
functionalities of Varlociraptor).