feat: methylation candidates (#394)
### Description

It is now possible to find methylation candidates in a Fasta File. New
subcommand is: --candidates <input> [output]
               - input: FASTA File with the DNA of interest
- output: Location of the resulting vcf file (optional, if not provided
stdOut)

### QC
<!-- Make sure that you can tick the boxes below. -->

* [ ] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [ ] The documentation at
https://github.com/varlociraptor/varlociraptor.github.io is updated in a
separate PR to reflect the changes or this is not necessary (e.g. if the
change does neither modify the calling grammar nor the behavior or
functionalities of Varlociraptor).

---------

Co-authored-by: Johannes Koester <johannes.koester@uk-essen.de>
Co-authored-by: Addimator <adpri100@hhhu.de>
Co-authored-by: Johannes Köster <johannes.koester@tu-dortmund.de>