fix: complete overhaul of homopolymer error model (to be more specific) and fix for strand bias model (to be more sensitive)
Improvements for artifact detection

fix: overhaul of homopolymer error model. Main idea: ignore fragments
that either support ref or alt without homopolymer error. For the rest,
rely on observed errors overall and compare which center of the error
distribution better fits the observed homopolymer errors, the variant or
the reference allele.

fix: calculate strand bias for cases where there is already a strong
bias in the ref fragments, if that bias can be reliably estimated
because there are enough supporting fragments.