fix: fix handling of N-bases in reads (so far, they could have yielded wrong likelihoods) (#444)
The fixed problem was very rare though, only mattering when the given
candidate variants are not derived from the data but e.g. known from
some database.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation at
https://github.com/varlociraptor/varlociraptor.github.io is updated in a
separate PR to reflect the changes or this is not necessary (e.g. if the
change does neither modify the calling grammar nor the behavior or
functionalities of Varlociraptor).

---------

Co-authored-by: Johannes Koester <johannes.koester@uni-due.de>