feat: support for per variant prior heterozygosity or somatic mutation rates (#448)
This allows to specify an INFO field in the candidates VCF that contains
(per alt allele) expected heterozygosity or somatic effective mutation
rate. For example this can be the AF field that is defined in the BCF
spec and presumably offered by clinvar, dbsnp or gnomad.

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation at
https://github.com/varlociraptor/varlociraptor.github.io is updated in a
separate PR to reflect the changes or this is not necessary (e.g. if the
change does neither modify the calling grammar nor the behavior or
functionalities of Varlociraptor).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **New Features**
- Introduced new fields for variant heterozygosity and somatic effective
mutation rate, enhancing genomic analysis capabilities.
- Updated processing pipeline to include new genetic metrics in variant
handling.
- Expanded command-line interface to capture additional variant-related
information.
- Enhanced functionality of the variant collection process to include
detailed information about heterozygosity and somatic mutation rates.
- Added methods to manage new metrics within the Prior struct and
related traits.

- **Bug Fixes**
- Added specific error handling for invalid variant prior entries,
improving clarity in error reporting.

- **Documentation**
- Enhanced metadata in output records to define new fields related to
variant heterozygosity and mutation rates.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->

---------

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>