fix: fixed handling of read position information when merging read pairs (#505)
### Description

<!--Add a description of your PR here-->

### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation at
https://github.com/varlociraptor/varlociraptor.github.io is updated in a
separate PR to reflect the changes or this is not necessary (e.g. if the
change does neither modify the calling grammar nor the behavior or
functionalities of Varlociraptor).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Refactor**
- Improved the logic for merging allele support information, resulting
in more accurate handling of read position data when assessing alternate
allele support. No changes to visible features or user interface.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->