fix: consider leading hardclips when calculating read position (#510)
### QC
<!-- Make sure that you can tick the boxes below. -->

* [x] The PR contains a test case for the changes or the changes are
already covered by an existing test case.
* [x] The documentation at
https://github.com/varlociraptor/varlociraptor.github.io is updated in a
separate PR to reflect the changes or this is not necessary (e.g. if the
change does neither modify the calling grammar nor the behavior or
functionalities of Varlociraptor).


<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->
## Summary by CodeRabbit

- **Bug Fixes**
- Corrected the calculation of read positions for variant reporting to
accurately include leading hard-clipped bases from sequencing reads.
This ensures that read positions now reflect the original coordinates,
improving the accuracy of variant annotations.
<!-- end of auto-generated comment: release notes by coderabbit.ai -->